Finished median and removed unsued attributes
Started and finished the median function, removed height and nodes
attributes and started trying to implement a new remove function that
doesn't depend on the search and has a recursive function that returns a
bool.

The height and nodes attributes weren't being used. I'm trying to
implement a remove function like the insert function, where the recursive
function returns a boolean, so I don't have to do a search to see if I
should or should not delete a node and update the values of nLeft and
nRight on each of the affected nodes.